server/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# ML Libraries
torch>=2.5.0
torchvision>=0.20.0
transformers==4.37.0
pillow==10.2.0

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6
email-validator==2.1.0

# Database
pymongo==4.6.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
